Adding aquaplanet initialization to init_atmosphere
This commit adds the initialization of an aquaplanet configuration to init_atmosphere. The aquaplanet is assigned test_case = 0. The main subroutine initializes the vertical model grid, reads in a sounding called sounding.input, interpolates the sounding to the vertical model grid, specifies an SST profile as a function of latitude, and defines all other necessary state variables. Changes are also made to make sure that if an user requests the aquaplanet configuration, the surface is set to an ocean surface (xland=2) and the surface albedo is set to a non-zero value.